Add renderMessage and renderInstruction to Error component
