+++ src/components/Camera/index.js
@@ -18,6 +18,7 @@ import classNames from 'classnames'
 import style from './style.css'
 import type { CameraPureType, CameraType, CameraActionType, CameraStateType} from './CameraTypes'
 import { checkIfWebcamPermissionGranted } from '../utils'
+import { parseTags } from '../utils/string'
 
 const UploadFallback = ({onUploadFallback, onFallbackClick, method, i18n}) => {
   const text = i18n && method ? i18n.t(`capture.${method}.help`) : ''
@@ -76,12 +77,11 @@ export class CameraPure extends React.Component<CameraPureType> {
               <div className={style.errorContainer}>
                 <Error
                   {...{i18n}}
-                  error={{ name: 'CAMERA_NOT_WORKING', type: 'error' }}
                   className={style.errorMessage}
-                  options={{
-                    parseInstructionTags: ({ text }) =>
-                      <span onClick={reload} className={style.errorLink}>{text}</span>,
-                  }}
+                  error={{ name: 'CAMERA_NOT_WORKING', type: 'error' }}
+                  renderInstruction={ str =>
+                    parseTags(str, ({ text }) => <span onClick={reload} className={style.errorLink}>{text}</span>)
+                  }
                   smaller
                 />
               </div> :
